conjoin post_migrate steps again
